converging some logic to single place
