+++ lib/src/commands/db_upgrade.dart
@@ -42,16 +42,13 @@ class CLIDatabaseUpgrade extends CLIDatabaseConnectingCommand {
     var schema = new Schema.empty();
     for (var migration in migrationFilesToGetToCurrent) {
       displayProgress("Replaying version ${versionNumberFromFile(migration)}");
-      schema = await executeUpgradeForFile(
-          projectDirectory, migration, schema, versionNumberFromFile(migration),
-          connectionInfo: _storeConnectionMap, dryRun: true);
+      schema = await schemaByApplyingMigrationFile(migration, schema);
     }
 
     for (var migration in migrationFilesToRun) {
       displayInfo("Applying version ${versionNumberFromFile(migration)}...");
-      schema = await executeUpgradeForFile(
-          projectDirectory, migration, schema, versionNumberFromFile(migration),
-          connectionInfo: _storeConnectionMap, dryRun: false);
+      schema =
+          await executeUpgradeForFile(migration, schema, _storeConnectionMap);
       displayProgress(
           "Applied version ${versionNumberFromFile(migration)} successfully.",
           color: CLIColor.green);
@@ -90,6 +87,69 @@ class CLIDatabaseUpgrade extends CLIDatabaseConnectingCommand {
     return [migrationFilesToGetToCurrent, migrationFilesToRun];
   }
 
+  Future<Schema> executeUpgradeForFile(
+      File file, Schema schema, Map<String, dynamic> connectionInfo) async {
+    var generator = new SourceGenerator(
+        (List<String> args, Map<String, dynamic> values) async {
+      hierarchicalLoggingEnabled = true;
+
+      PostgreSQLPersistentStore.logger.level = Level.ALL;
+      PostgreSQLPersistentStore.logger.onRecord
+          .listen((r) => print("\t${r.message}"));
+
+      var inputSchema =
+          new Schema.fromMap(values["schema"] as Map<String, dynamic>);
+      var dbInfo = values["dbInfo"];
+
+      PersistentStore store;
+      if (dbInfo != null && dbInfo["flavor"] == "postgres") {
+        store = new PostgreSQLPersistentStore.fromConnectionInfo(
+            dbInfo["username"],
+            dbInfo["password"],
+            dbInfo["host"],
+            dbInfo["port"],
+            dbInfo["databaseName"],
+            timeZone: dbInfo["timeZone"]);
+      }
+
+      var versionNumber = int.parse(args.first);
+      var migrationClassMirror = currentMirrorSystem()
+              .isolate
+              .rootLibrary
+              .declarations
+              .values
+              .firstWhere((dm) =>
+                  dm is ClassMirror && dm.isSubclassOf(reflectClass(Migration)))
+          as ClassMirror;
+
+      var migrationInstance = migrationClassMirror
+          .newInstance(new Symbol(''), []).reflectee as Migration;
+      migrationInstance.database = new SchemaBuilder(store, inputSchema);
+
+      await migrationInstance.upgrade();
+
+      await migrationInstance.store
+          .upgrade(versionNumber, migrationInstance.database.commands);
+      await migrationInstance.seed();
+      await migrationInstance.database.store.close();
+
+      return migrationInstance.currentSchema.asMap();
+    }, imports: [
+      "dart:async",
+      "package:aqueduct/aqueduct.dart",
+      "package:logging/logging.dart",
+      "dart:isolate",
+      "dart:mirrors"
+    ], additionalContents: file.readAsStringSync());
+
+    var executor = new IsolateExecutor(
+        generator, ["${versionNumberFromFile(file)}"],
+        message: {"schema": schema.asMap(), "dbInfo": connectionInfo});
+
+    var schemaMap = await executor.execute(projectDirectory.uri);
+    return new Schema.fromMap(schemaMap as Map<String, dynamic>);
+  }
+
   Map<String, dynamic> get _storeConnectionMap {
     var s = persistentStore;
     if (s is PostgreSQLPersistentStore) {
@@ -106,77 +166,4 @@ class CLIDatabaseUpgrade extends CLIDatabaseConnectingCommand {
 
     return null;
   }
-}
-
-Future<Schema> executeUpgradeForFile(
-    Directory projectDirectory, File file, Schema schema, int versionNumber,
-    {Map<String, dynamic> connectionInfo, bool dryRun: false, bool enableLogging: true}) async {
-  var generator = new SourceGenerator(
-      (List<String> args, Map<String, dynamic> values) async {
-    if (values["enableLogging"]) {
-      hierarchicalLoggingEnabled = true;
-
-      PostgreSQLPersistentStore.logger.level = Level.ALL;
-      PostgreSQLPersistentStore.logger.onRecord
-        .listen((r) => print("\t${r.message}"));
-    }
-
-    var inputSchema =
-        new Schema.fromMap(values["schema"] as Map<String, dynamic>);
-    var dbInfo = values["dbInfo"];
-    var dryRun = values["dryRun"];
-
-    PersistentStore store;
-    if (dbInfo != null && dbInfo["flavor"] == "postgres") {
-      store = new PostgreSQLPersistentStore.fromConnectionInfo(
-          dbInfo["username"],
-          dbInfo["password"],
-          dbInfo["host"],
-          dbInfo["port"],
-          dbInfo["databaseName"],
-          timeZone: dbInfo["timeZone"]);
-    }
-
-    var versionNumber = int.parse(args.first);
-    var migrationClassMirror = currentMirrorSystem()
-            .isolate
-            .rootLibrary
-            .declarations
-            .values
-            .firstWhere((dm) =>
-                dm is ClassMirror && dm.isSubclassOf(reflectClass(Migration)))
-        as ClassMirror;
-    var migrationInstance = migrationClassMirror
-        .newInstance(new Symbol(''), []).reflectee as Migration;
-    migrationInstance.database = new SchemaBuilder(store, inputSchema);
-
-    await migrationInstance.upgrade();
-
-    if (!dryRun) {
-      await migrationInstance.store
-          .upgrade(versionNumber, migrationInstance.database.commands);
-      await migrationInstance.seed();
-      await migrationInstance.database.store.close();
-    }
-
-    return migrationInstance.currentSchema.asMap();
-  }, imports: [
-    "dart:async",
-    "package:aqueduct/aqueduct.dart",
-    "package:logging/logging.dart",
-    "dart:isolate",
-    "dart:mirrors"
-  ], additionalContents: file.readAsStringSync());
-
-  var executor = new IsolateExecutor(generator, [
-    "${versionNumber}"
-  ], message: {
-    "dryRun": dryRun,
-    "schema": schema.asMap(),
-    "dbInfo": connectionInfo,
-    "enableLogging": enableLogging
-  });
-
-  var schemaMap = await executor.execute(projectDirectory.uri);
-  return new Schema.fromMap(schemaMap as Map<String, dynamic>);
 }
