+++ lib/src/commands/db_validate.dart
@@ -1,6 +1,4 @@
 import 'dart:async';
-import 'dart:io';
-import 'dart:mirrors';
 
 import '../db/db.dart';
 import '../utilities/source_generator.dart';
@@ -39,7 +37,7 @@ class CLIDatabaseValidate extends CLICommand
     var schemaFromMigrationFiles = new Schema.empty();
     for (var migrationFile in migrationFiles) {
       schemaFromMigrationFiles =
-          await schemaByApplyingMigrationFile(projectDirectory, migrationFile, schemaFromMigrationFiles, versionNumberFromFile(migrationFile));
+          await schemaByApplyingMigrationFile(migrationFile, schemaFromMigrationFiles);
     }
 
     var differences = currentSchema.differenceFrom(schemaFromMigrationFiles);
@@ -70,44 +68,3 @@ class CLIDatabaseValidate extends CLICommand
 }
 
 
-Future<Schema> schemaByApplyingMigrationFile(
-    Directory projectDirectory, File migrationFile, Schema baseSchema, int versionNumber) async {
-  var sourceFunction =
-      (List<String> args, Map<String, dynamic> values) async {
-    var inputSchema =
-    new Schema.fromMap(values["schema"] as Map<String, dynamic>);
-
-    var versionNumber = int.parse(args.first);
-    var migrationClassMirror = currentMirrorSystem()
-        .isolate
-        .rootLibrary
-        .declarations
-        .values
-        .firstWhere((dm) =>
-    dm is ClassMirror && dm.isSubclassOf(reflectClass(Migration)))
-    as ClassMirror;
-
-    var migrationInstance = migrationClassMirror
-        .newInstance(new Symbol(''), []).reflectee as Migration;
-    migrationInstance.database = new SchemaBuilder(null, inputSchema);
-
-    await migrationInstance.upgrade();
-
-    return migrationInstance.currentSchema.asMap();
-  };
-
-  var generator = new SourceGenerator(sourceFunction,
-      imports: [
-        "dart:async",
-        "package:aqueduct/aqueduct.dart",
-        "dart:isolate",
-        "dart:mirrors"
-      ],
-      additionalContents: migrationFile.readAsStringSync());
-
-  var schemaMap = await IsolateExecutor.executeSource(generator,
-      ["${versionNumber}"], projectDirectory.uri,
-      message: {"schema": baseSchema.asMap()});
-
-  return new Schema.fromMap(schemaMap as Map<String, dynamic>);
-}
